+++ compute/runtime/init.go
@@ -4,8 +4,8 @@ import (
 	"os"
 
 	"github.com/carv-ics-forth/hpk/compute"
+	"github.com/carv-ics-forth/hpk/compute/endpoint"
 	"github.com/carv-ics-forth/hpk/compute/image"
-	"github.com/carv-ics-forth/hpk/compute/paths"
 	"github.com/pkg/errors"
 )
 
@@ -15,20 +15,20 @@ var (
 )
 
 func Initialize() error {
-	compute.HPK = paths.HPK(compute.Environment.WorkingDirectory)
+	compute.HPK = endpoint.HPK(compute.Environment.WorkingDirectory)
 
 	// create the ~/.hpk directory, if it does not exist.
-	if err := os.MkdirAll(compute.HPK.String(), paths.PodGlobalDirectoryPermissions); err != nil {
+	if err := os.MkdirAll(compute.HPK.String(), endpoint.PodGlobalDirectoryPermissions); err != nil {
 		return errors.Wrapf(err, "Failed to create RuntimeDir '%s'", compute.HPK.String())
 	}
 
 	// create the ~/.hpk/image directory, if it does not exist.
-	if err := os.MkdirAll(compute.HPK.ImageDir(), paths.PodGlobalDirectoryPermissions); err != nil {
+	if err := os.MkdirAll(compute.HPK.ImageDir(), endpoint.PodGlobalDirectoryPermissions); err != nil {
 		return errors.Wrapf(err, "Failed to create ImageDir '%s'", compute.HPK.ImageDir())
 	}
 
 	// create the ~/.hpk/corrupted directory, if it does not exist.
-	if err := os.MkdirAll(compute.HPK.CorruptedDir(), paths.PodGlobalDirectoryPermissions); err != nil {
+	if err := os.MkdirAll(compute.HPK.CorruptedDir(), endpoint.PodGlobalDirectoryPermissions); err != nil {
 		return errors.Wrapf(err, "Failed to create CorruptedDir '%s'", compute.HPK.CorruptedDir())
 	}