+++ examples/apps/opengadget3/install.sh
@@ -16,9 +16,18 @@ set -eu
 helm repo add frisbee https://carv-ics-forth.github.io/frisbee/charts
 helm repo update
 
+# Install cert-manager (if not already installled)
+helm status -n cert-manager cert-manager|| \
+helm install  cert-manager jetstack/cert-manager  --wait\
+       --create-namespace -n cert-manager \
+       --version v1.12.0 \
+       --set installCRDs=true \
+       --set webhook.securePort=10260
+
 # Install Frisbee (if not already installed)
 helm status frisbee -n frisbee ||  \
-helm install --atomic frisbee frisbee/platform --wait -n frisbee \
+helm install --atomic frisbee frisbee/platform --wait \
+    --create-namespace -n frisbee\
     --set chaos-mesh.enabled=false
 
 # Install Frisbee templates.
@@ -32,7 +41,7 @@ kubectl apply -f manifest.yaml -n "${TEST_NAMESPACE}"
 
 
 # Access Grafana from your workstation.
-echo -e "Waiting for Grafan to become ready ..."
+echo -e "Waiting for Grafana to become ready ..."
 sleep 10
 kubectl wait pods -n  "${TEST_NAMESPACE}" grafana --for condition=Ready --timeout=120s
 
@@ -40,7 +49,8 @@ CONTAINER_IP=$(kubectl get endpoints grafana -n "${TEST_NAMESPACE}" -o=jsonpath=
 CONTAINER_PORT=$(kubectl get endpoints grafana -n "${TEST_NAMESPACE}" -o=jsonpath='{.subsets[0].ports[0].port}')
 COMPUTE_NODE=$(kubectl get pods grafana -n "${TEST_NAMESPACE}" -o=jsonpath='{.metadata.annotations.pod\.hpk\/id}' | cut -d '/' -f 3 |  xargs -I {} squeue --job {} --noheader -o '%B')
 
-echo -e "To access Grafana from your workstation (http://127.0.0.1:${CONTAINER_PORT}):\n"
-echo ssh -t "<LOGIN_NODE>" -L ${CONTAINER_PORT}:127.0.0.1:${CONTAINER_PORT}  \
+echo "****************"
+echo "Access Grafana from your workstation (http://127.0.0.1:${CONTAINER_PORT}):"
+echo ssh -t "\${LOGIN_NODE}" -L ${CONTAINER_PORT}:127.0.0.1:${CONTAINER_PORT}  \
      ssh -L ${CONTAINER_PORT}:${CONTAINER_IP}:${CONTAINER_PORT} ${COMPUTE_NODE}
-
+echo "****************"